+++ mobile/android/android-components/components/feature/prompts/src/main/java/mozilla/components/feature/prompts/concept/SelectablePromptView.kt
@@ -31,15 +31,6 @@ interface SelectablePromptView<T> {
     fun asView(): View = (this as View)
 
     /**
-     * Tries to inflate the view if needed.
-     *
-     * See: https://github.com/mozilla-mobile/android-components/issues/5491
-     *
-     * @return true if the inflation was completed, false if the view was already inflated.
-     */
-    fun tryInflate(): Boolean
-
-    /**
      * Interface to allow a class to listen to the option selection prompt events.
      */
     interface Listener<in T> {